requirements.txt
gunicorn
uvicorn
auto_mix_prep==0.2.0
config==0.5.1
openai~=1.16.2
pydantic==2.6.4
pydub==0.25.1
pyloudnorm==0.1.1
python-dotenv==1.0.1
requests_oauthlib==2.0.0
soundfile==0.12.1
decorator==4.4.2
fastapi==0.110.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
openai==1.16.2
pydantic==2.6.4
python-dotenv==1.0.1
requests-oauthlib==2.0.0
uvicorn==0.29.0
requests~=2.31.0
pathlib~=1.0.1
opencv-python~=4.9.0.80
pytest
Pillow
google-cloud-resource-manager
google-cloud-pubsub
kubernetes
APScheduler